+++ waspc/src/Wasp/Generator.hs
@@ -33,7 +33,7 @@ import qualified Wasp.Generator.Test
 import Wasp.Generator.Valid (validateExternalConfigsWithAppSpec)
 import qualified Wasp.Generator.WaspInfo as WaspInfo
 import Wasp.Generator.WaspLibs (genWaspLibs)
-import Wasp.Generator.WriteFileDrafts (synchronizeFileDraftsWithDisk)
+import Wasp.Generator.WriteFileDrafts (GeneratedAppPathChange, synchronizeFileDraftsWithDisk)
 import Wasp.Message (SendMessage)
 import Wasp.Util ((<++>))
 
@@ -45,20 +45,20 @@ import Wasp.Util ((<++>))
 --   NOTE(martin): What if there is already smth in the dstDir? It is probably best
 --     if we clean it up first? But we don't want this to end up with us deleting stuff
 --     from user's machine. Maybe we just overwrite and we are good?
-writeWebAppCode :: AppSpec -> Path' Abs (Dir GeneratedAppDir) -> SendMessage -> IO ([GeneratorWarning], [GeneratorError])
+writeWebAppCode :: AppSpec -> Path' Abs (Dir GeneratedAppDir) -> SendMessage -> IO ([GeneratorWarning], [GeneratorError], [GeneratedAppPathChange])
 writeWebAppCode spec dstDir sendMessage = do
   case validateExternalConfigsWithAppSpec spec of
-    validationErrors@(_ : _) -> return ([], validationErrors)
+    validationErrors@(_ : _) -> return ([], validationErrors, [])
     [] -> do
       let (generatorWarnings, generatorResult) = runGenerator $ genApp spec
 
       case generatorResult of
-        Left generatorErrors -> return (generatorWarnings, toList generatorErrors)
+        Left generatorErrors -> return (generatorWarnings, toList generatorErrors, [])
         Right fileDrafts -> do
-          synchronizeFileDraftsWithDisk dstDir fileDrafts
+          generatedAppPathChanges <- synchronizeFileDraftsWithDisk dstDir fileDrafts
           WaspInfo.persist dstDir $ AS.buildType spec
           (setupGeneratorWarnings, setupGeneratorErrors) <- runSetup spec dstDir sendMessage
-          return (generatorWarnings ++ setupGeneratorWarnings, setupGeneratorErrors)
+          return (generatorWarnings ++ setupGeneratorWarnings, setupGeneratorErrors, generatedAppPathChanges)
 
 genApp :: AppSpec -> Generator [FileDraft]
 genApp spec = do